Refactor `AzimuthBaseSettings` outside of `config.py` (#557)
